Remove .hpp files
