+++ include/cachemere/detail/traits.h
@@ -1,6 +1,8 @@
 #pragma once
 
-#include "cachemere/item.h"
+#include <boost/hana.hpp>
+
+#include <cachemere/item.h>
 
 namespace cachemere::detail::traits {
