+++ include/cachemere/measurement.h
@@ -1,9 +1,32 @@
 #pragma once
 
-#include <cstdint>
+#include <memory>
 
 /// @brief Utilities for measuring cached items.
 namespace cachemere::measurement {
+
+namespace detail {
+
+// Utility functions to allow passing both object and pointers to objects
+// to the measurement functors.
+template<typename T> inline const T& deref_maybe(const T& object)
+{
+    return object;
+}
+template<typename T> inline const T& deref_maybe(T* object)
+{
+    return *object;
+}
+template<typename T> inline const T& deref_maybe(const std::shared_ptr<T>& object)
+{
+    return *object;
+}
+template<typename T> inline const T& deref_maybe(const std::unique_ptr<T>& object)
+{
+    return *object;
+}
+
+}  // namespace detail
 
 /// @brief Get the size of an object via a user-defined `size()` method.
 template<typename T> struct Size {
@@ -24,6 +47,30 @@ private:
     [[nodiscard]] size_t round_up(const size_t value) const;
 };
 
-}  // namespace cachemere::measurement
 
-#include "measurement.hpp"
+template<typename T> template<typename V> size_t Size<T>::operator()(const V& object) const
+{
+    return detail::deref_maybe(object).size();
+}
+
+template<typename T> template<typename V> size_t SizeOf<T>::operator()(const V& /* object */) const
+{
+    return sizeof(T);
+}
+
+template<typename T> template<typename V> size_t CapacityDynamicallyAllocated<T>::operator()(const V& object) const
+{
+    size_t capacity = detail::deref_maybe(object).capacity();
+    if (capacity < 1024) {
+        capacity = std::max(static_cast<size_t>(16), round_up(capacity));
+    }
+
+    return capacity;
+}
+
+template<typename T> size_t CapacityDynamicallyAllocated<T>::round_up(const size_t value) const
+{
+    return ((value + sizeof(void*) - 1) / sizeof(void*)) * sizeof(void*);
+}
+
+}  // namespace cachemere::measurement
