+++ include/cachemere/policy/detail/hash_mixer.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <cstdint>
 #include <random>
 
 namespace cachemere::policy::detail {
@@ -27,6 +26,17 @@ private:
     size_t           m_value_range;
 };
 
-}  // namespace cachemere::policy::detail
+template<typename Key, typename KeyHash>
+HashMixer<Key, KeyHash>::HashMixer(const Key& key, size_t value_range)
+ : KeyHash{},
+   m_rng{static_cast<std::minstd_rand::result_type>(KeyHash::operator()(key))},
+   m_value_range{value_range}
+{
+}
 
-#include "hash_mixer.hpp"
+template<typename Key, typename KeyHash> size_t HashMixer<Key, KeyHash>::operator()()
+{
+    return m_rng() % m_value_range;
+}
+
+}  // namespace cachemere::policy::detail
