+++ include/cachemere/policy/insertion_always.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <iostream>
-
 namespace cachemere::policy {
 
 /// @brief Simplest insertion policy. Always accepts insertions.
@@ -28,6 +26,19 @@ public:
     bool should_replace(const Key& victim, const Key& candidate);
 };
 
-}  // namespace cachemere::policy
+template<typename Key, typename KeyHash, typename Value> void InsertionAlways<Key, KeyHash, Value>::clear()
+{
+}
 
-#include "insertion_always.hpp"
+template<typename Key, typename KeyHash, typename Value> bool InsertionAlways<Key, KeyHash, Value>::should_add(const Key& /* key */)
+{
+    return true;
+}
+
+template<typename Key, typename KeyHash, typename Value>
+bool InsertionAlways<Key, KeyHash, Value>::should_replace(const Key& /* key */, const Key& /* candidate */)
+{
+    return true;
+}
+
+}  // namespace cachemere::policy
